% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dst.R
\name{is.tbl_dst}
\alias{is.tbl_dst}
\title{is.tbl_dst}
\usage{
is.tbl_dst(x)
}
\arguments{
\item{x}{[tbl_dst] output.}
}
\value{
[logical] TRUE/FALSE.
}
\description{
\code{is.tbl_dst} renames to pretty names
}
\examples{

\dontrun{
use_names(x)
}
}
