% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dst.R
\name{use_bulk_download}
\alias{use_bulk_download}
\title{Bulk download}
\usage{
use_bulk_download(x, bulk = TRUE)
}
\arguments{
\item{x}{[tbl_dst] output.}

\item{bulk}{[logical] output.}
}
\value{
a [tbl_dst] object.
}
\description{
\code{use_names} renames to pretty names
}
