% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dst.R
\name{use_labels}
\alias{use_labels}
\title{Encode with labels}
\usage{
use_labels(x, value = TRUE)
}
\arguments{
\item{x}{[tbl_dst] output.}

\item{value}{[logical] TRUE/FALSE}
}
\value{
a [tbl_dst] object.
}
\description{
\code{use_names} renames to pretty names
}
