% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dst.R
\name{use_long_names}
\alias{use_long_names}
\title{Rename}
\usage{
use_long_names(x, value = TRUE)
}
\arguments{
\item{x}{[tbl_dst] output.}

\item{value}{[logical] TRUE/FALSE}
}
\value{
a [tbl_dst] object.
}
\description{
\code{use_long_names} renames to pretty names
}
\examples{

\dontrun{
use_names(x)
}
}
