% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dst.R
\name{use_names}
\alias{use_names}
\title{Rename}
\usage{
use_names(x)
}
\arguments{
\item{x}{[tbl_dst] output.}
}
\value{
a [tbl_dst] object.
}
\description{
\code{use_names} renames to pretty names
}
\examples{

\dontrun{
use_names(x)
}
}
